requirements.txt
# Core dependencies
streamlit>=1.20.0
pandas>=1.3.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
pymysql>=1.0.2
cryptography>=36.0.0  # Required for MySQL authentication
nest_asyncio>=1.5.5  # For compatibility with the working app

# Data processing and export
pyarrow>=7.0.0
fastparquet>=0.8.0
openpyxl>=3.0.9  # For Excel export

# Compression
zipfile36>=0.1.3  # If using Python < 3.8, otherwise built-in